Initial commit; migrated from archive project
I've been maintaining this in the archive but now doing enough with it
that I think it's time to give it its own repo. This is the latest
commit from the project at the time of the migration:
https://github.com/tylerbrazier/archive/commit/056f5776230af987338c2b8490305d9bb0623183
I'll probably remove the code from archive until I'm no longer using it.